Split up tests into separate files and simplify
