+++ data-in-pipeline/tests/transform/test_navigator_family.py
@@ -8,14 +8,12 @@ from data_in_models.models import (
     Item,
     Label,
     LabelRelationship,
-    LabelWithoutDocumentRelationships,
 )
 
 from app.extract.connectors import (
     NavigatorFamily,
 )
-from app.models import Identified, NavigatorConcept
-from app.transform.models import CouldNotTransform
+from app.models import Identified
 from app.transform.navigator_family import transform_navigator_family
 from tests.factories import (
     NavigatorCollectionFactory,
@@ -30,17 +28,6 @@ from tests.transform.assertions import assert_model_list_equality
 
 
 def _cclw_corpus():
-    return NavigatorCorpusFactory.build(
-        import_id="CCLW.corpus.i00000001.n0000",
-        corpus_type=NavigatorCorpusTypeFactory.build(name="corpus_type"),
-        organisation=NavigatorOrganisationFactory.build(id=1, name="CCLW"),
-        attribution_url="testurl.org",
-        corpus_text="Test corpus",
-        corpus_image_url="corpus_image.png",
-    )
-
-
-def _cclw_laws_and_policies_corpus():
     return NavigatorCorpusFactory.build(
         import_id="CCLW.corpus.i00000001.n0000",
         corpus_type=NavigatorCorpusTypeFactory.build(name="Laws and Policies"),
@@ -98,7 +85,7 @@ def navigator_family_with_single_matching_document() -> Identified[NavigatorFami
             title="Matching title on family and document and collection",
             summary="Family summary",
             category="EXECUTIVE",
-            corpus=_cclw_laws_and_policies_corpus(),
+            corpus=_cclw_corpus(),
             last_updated_date="2020-01-0100:00:00Z",
             published_date="2020-01-0100:00:00Z",
             documents=[
@@ -170,429 +157,6 @@ def navigator_family_with_no_matching_transformations() -> Identified[NavigatorF
             events=[],
             collections=[],
             geographies=[],
-        ),
-    )
-
-
-@pytest.fixture
-def navigator_family_with_litigation_corpus_type() -> Identified[NavigatorFamily]:
-    decision_date = datetime.datetime(2020, 1, 1)
-    return Identified(
-        id="family",
-        source="navigator_family",
-        data=NavigatorFamilyFactory.build(
-            import_id="family",
-            title="Litigation family",
-            summary="Family summary",
-            category="LITIGATION",
-            last_updated_date="2020-01-0100:00:00Z",
-            published_date="2020-01-0100:00:00Z",
-            corpus=NavigatorCorpusFactory.build(
-                import_id="Academic.corpus.Litigation.n0000",
-                corpus_type=NavigatorCorpusTypeFactory.build(name="Litigation"),
-                organisation=NavigatorOrganisationFactory.build(id=1, name="CCLW"),
-                attribution_url="testurl.org",
-                corpus_text="Test corpus",
-                corpus_image_url=None,
-            ),
-            documents=[
-                NavigatorDocumentFactory.build(
-                    import_id="document",
-                    title="Litigation family document",
-                    cdn_object=None,
-                    source_url=None,
-                    slug="litigation-document-slug",
-                    events=[
-                        NavigatorEventFactory.build(
-                            import_id="123",
-                            event_type="Decision",
-                            date=decision_date,
-                            valid_metadata={
-                                "event_type": ["Decision"],
-                                "datetime_event_name": ["Decision"],
-                            },
-                        )
-                    ],
-                    variant="Original language",
-                    md5_sum="aaaaa11111bbbbb",
-                    languages=[],
-                    document_status="PUBLISHED",
-                ),
-                NavigatorDocumentFactory.build(
-                    import_id="1.2.3.placeholder",
-                    title="Placeholder litigation family document",
-                    cdn_object=None,
-                    source_url=None,
-                    slug="placeholder-document-slug",
-                    events=[],
-                    variant=None,
-                    md5_sum="aaaaa11111bbbbb",
-                    languages=[],
-                    document_status="PUBLISHED",
-                ),
-            ],
-            events=[
-                NavigatorEventFactory.build(
-                    import_id="123",
-                    event_type="Decision",
-                    date=decision_date,
-                    valid_metadata={
-                        "event_type": ["Decision"],
-                        "datetime_event_name": ["Decision"],
-                    },
-                ),
-                NavigatorEventFactory.build(
-                    import_id="456",
-                    event_type="Filing Year For Action",
-                    date=decision_date,
-                    valid_metadata={
-                        "event_type": ["Filing Year For Action"],
-                        "datetime_event_name": ["Filing Year For Action"],
-                    },
-                ),
-            ],
-            collections=[],
-            geographies=[],
-            slug="litigation-family-slug",
-            metadata={
-                "case_number": ["CASE-NUMBER 123"],
-                "core_object": ["Core Object 123"],
-            },
-            concepts=[],
-        ),
-    )
-
-
-@pytest.fixture
-def navigator_family_with_litigation_concepts() -> Identified[NavigatorFamily]:
-    decision_date = datetime.datetime(2020, 1, 1)
-    return Identified(
-        id="family",
-        source="navigator_family",
-        data=NavigatorFamilyFactory.build(
-            import_id="family",
-            title="Litigation family",
-            summary="Family summary",
-            category="LITIGATION",
-            last_updated_date="2020-01-0100:00:00Z",
-            published_date="2020-01-0100:00:00Z",
-            corpus=NavigatorCorpusFactory.build(
-                import_id="Academic.corpus.Litigation.n0000",
-                corpus_type=NavigatorCorpusTypeFactory.build(name="Litigation"),
-                organisation=NavigatorOrganisationFactory.build(id=1, name="Sabin"),
-                attribution_url="testurl.org",
-                corpus_text="Test corpus",
-                corpus_image_url=None,
-            ),
-            documents=[
-                NavigatorDocumentFactory.build(
-                    import_id="document",
-                    title="Litigation family document",
-                    cdn_object=None,
-                    source_url=None,
-                    slug="litigation-document-slug",
-                    events=[
-                        NavigatorEventFactory.build(
-                            import_id="123",
-                            event_type="Decision",
-                            date=decision_date,
-                            valid_metadata={
-                                "event_type": ["Decision"],
-                                "datetime_event_name": ["Decision"],
-                            },
-                        )
-                    ],
-                    variant="Original language",
-                    md5_sum="aaaaa11111bbbbb",
-                    languages=[],
-                    document_status="PUBLISHED",
-                ),
-            ],
-            events=[
-                NavigatorEventFactory.build(
-                    import_id="123",
-                    event_type="Decision",
-                    date=decision_date,
-                    valid_metadata={
-                        "event_type": ["Decision"],
-                        "datetime_event_name": ["Decision"],
-                    },
-                ),
-            ],
-            collections=[],
-            geographies=[],
-            slug="litigation-family-slug",
-            concepts=[
-                NavigatorConcept(
-                    id="High Court of Justice",
-                    ids=[],
-                    type="legal_entity",
-                    relation="jurisdiction",
-                    preferred_label="High Court of Justice",
-                    subconcept_of_labels=["England and Wales"],
-                ),
-                NavigatorConcept(
-                    id="High Court of Justice (Administrative Court)",
-                    ids=[],
-                    type="legal_entity",
-                    relation="jurisdiction",
-                    preferred_label="High Court of Justice (Administrative Court)",
-                    subconcept_of_labels=["High Court of Justice"],
-                ),
-                NavigatorConcept(
-                    id="England and Wales",
-                    ids=[],
-                    type="legal_entity",
-                    relation="jurisdiction",
-                    preferred_label="England and Wales",
-                    subconcept_of_labels=[],
-                ),
-            ],
-        ),
-    )
-
-
-@pytest.fixture
-def navigator_family_with_litigation_concept_missing_parent() -> (
-    Identified[NavigatorFamily]
-):
-    decision_date = datetime.datetime(2020, 1, 1)
-    return Identified(
-        id="family",
-        source="navigator_family",
-        data=NavigatorFamilyFactory.build(
-            import_id="family",
-            title="Litigation family",
-            summary="Family summary",
-            category="LITIGATION",
-            last_updated_date="2020-01-0100:00:00Z",
-            published_date="2020-01-0100:00:00Z",
-            corpus=NavigatorCorpusFactory.build(
-                import_id="Academic.corpus.Litigation.n0000",
-                corpus_type=NavigatorCorpusTypeFactory.build(name="Litigation"),
-                organisation=NavigatorOrganisationFactory.build(id=1, name="Sabin"),
-                attribution_url="testurl.org",
-                corpus_text="Test corpus",
-                corpus_image_url=None,
-            ),
-            documents=[
-                NavigatorDocumentFactory.build(
-                    import_id="document",
-                    title="Litigation family document",
-                    slug="litigation-document-slug",
-                    events=[
-                        NavigatorEventFactory.build(
-                            import_id="123",
-                            event_type="Decision",
-                            date=decision_date,
-                            valid_metadata={
-                                "event_type": ["Decision"],
-                                "datetime_event_name": ["Decision"],
-                            },
-                        )
-                    ],
-                    variant="Original language",
-                    md5_sum="aaaaa11111bbbbb",
-                    languages=[],
-                    document_status="PUBLISHED",
-                ),
-            ],
-            events=[],
-            collections=[],
-            geographies=[],
-            slug="litigation-family-slug",
-            concepts=[
-                NavigatorConcept(
-                    id="High Court of Justice",
-                    ids=[],
-                    type="legal_entity",
-                    relation="jurisdiction",
-                    preferred_label="High Court of Justice",
-                    subconcept_of_labels=["Missing Parent Label"],
-                ),
-            ],
-        ),
-    )
-
-
-def _mcf_events():
-    base_date = datetime.datetime(2020, 1, 1)
-    return [
-        NavigatorEventFactory.build(
-            import_id="concept_approved",
-            event_type="Concept Approved",
-            date=base_date,
-            valid_metadata={
-                "event_type": ["Concept Approved"],
-                "datetime_event_name": ["Project Approved"],
-            },
-        ),
-        NavigatorEventFactory.build(
-            import_id="project_approved",
-            event_type="Project Approved",
-            date=base_date,
-            valid_metadata={
-                "event_type": ["Project Approved"],
-                "datetime_event_name": ["Project Approved"],
-            },
-        ),
-        NavigatorEventFactory.build(
-            import_id="under_implementation",
-            event_type="Under Implementation",
-            date=base_date,
-            valid_metadata={
-                "event_type": ["Under Implementation"],
-                "datetime_event_name": ["Project Approved"],
-            },
-        ),
-        NavigatorEventFactory.build(
-            import_id="project_completed",
-            event_type="Project Completed",
-            date=base_date,
-            valid_metadata={
-                "event_type": ["Project Completed"],
-                "datetime_event_name": ["Project Approved"],
-            },
-        ),
-        NavigatorEventFactory.build(
-            import_id="cancelled",
-            event_type="Cancelled",
-            date=base_date,
-            valid_metadata={
-                "event_type": ["Cancelled"],
-                "datetime_event_name": ["Project Approved"],
-            },
-        ),
-    ]
-
-
-@pytest.fixture
-def navigator_family_multilateral_climate_fund_project() -> Identified[NavigatorFamily]:
-    return Identified(
-        id="family",
-        source="navigator_family",
-        data=NavigatorFamilyFactory.build(
-            import_id="family",
-            title="Multilateral climate fund project",
-            summary="Family summary",
-            category="MCF",
-            last_updated_date=None,
-            published_date=None,
-            corpus=NavigatorCorpusFactory.build(
-                import_id="MCF.corpus.AF.n0000",
-                corpus_type=NavigatorCorpusTypeFactory.build(name="AF"),
-                organisation=NavigatorOrganisationFactory.build(id=1, name="CCLW"),
-                attribution_url="testurl.org",
-                corpus_text="Test corpus",
-                corpus_image_url=None,
-            ),
-            documents=[
-                NavigatorDocumentFactory.build(
-                    import_id="document_1",
-                    title="Multilateral climate fund project document",
-                    cdn_object=None,
-                    source_url=None,
-                    slug="document-1-slug",
-                    events=[],
-                    variant="Original language",
-                    md5_sum="aaaaa11111bbbbb",
-                    language="eng",
-                    languages=["eng"],
-                    document_status="PUBLISHED",
-                ),
-                NavigatorDocumentFactory.build(
-                    import_id="document_2",
-                    title="Project document",
-                    cdn_object=None,
-                    source_url=None,
-                    slug="document-2-slug",
-                    events=[],
-                    variant="Original language",
-                    md5_sum="aaaaa11111bbbbb",
-                    language="eng",
-                    languages=["eng"],
-                    document_status="PUBLISHED",
-                ),
-            ],
-            events=_mcf_events(),
-            collections=[],
-            geographies=[],
-            metadata={
-                "region": ["Europe & Central Asia"],
-                "sector": ["Public Sector"],
-                "status": ["Under Implementation"],
-                "project_id": ["XACTMK002A"],
-                "external_id": [],
-                "project_url": ["https://www.cif.org/projects"],
-                "implementing_agency": ["International Bank for Reconstruction"],
-                "project_value_fund_spend": ["250000"],
-                "project_value_co_financing": ["100000"],
-                "approved_ref": ["XACTMK002A"],
-            },
-            slug="mcf-family-slug",
-        ),
-    )
-
-
-@pytest.fixture
-def navigator_family_multilateral_climate_fund_guidance() -> (
-    Identified[NavigatorFamily]
-):
-    return Identified(
-        id="family",
-        source="navigator_family",
-        data=NavigatorFamilyFactory.build(
-            import_id="family",
-            title="Multilateral climate fund guidance",
-            summary="Family summary",
-            category="REPORTS",
-            last_updated_date=None,
-            published_date=None,
-            corpus=NavigatorCorpusFactory.build(
-                import_id="MCF.corpus.AF.Guidance",
-                corpus_type=NavigatorCorpusTypeFactory.build(name="AF"),
-                organisation=NavigatorOrganisationFactory.build(id=1, name="CCLW"),
-                attribution_url="testurl.org",
-                corpus_text="Test corpus",
-                corpus_image_url=None,
-            ),
-            documents=[],
-            events=[],
-            collections=[],
-            geographies=[],
-            metadata={},
-            slug="mcf-guidance-family-slug",
-        ),
-    )
-
-
-@pytest.fixture
-def navigator_family_with_duplicate_legal_case() -> Identified[NavigatorFamily]:
-    return Identified(
-        id="family",
-        source="navigator_family",
-        data=NavigatorFamilyFactory.build(
-            import_id="family",
-            title="Litigation family",
-            category="LITIGATION",
-            last_updated_date=None,
-            published_date=None,
-            summary="Family summary",
-            corpus=NavigatorCorpusFactory.build(
-                import_id="Academic.corpus.Litigation.n0000",
-                corpus_type=NavigatorCorpusTypeFactory.build(name="Litigation"),
-                organisation=NavigatorOrganisationFactory.build(
-                    id=1, name="Sabin Center for Climate Change Law"
-                ),
-                attribution_url="testurl.org",
-                corpus_text="Test corpus",
-                corpus_image_url=None,
-            ),
-            documents=[],
-            events=[],
-            collections=[],
-            geographies=[],
-            concepts=[],
         ),
     )
 
@@ -1330,897 +894,6 @@ def test_transform_navigator_family_with_laws_and_policies_corpus_type(
     )
 
 
-def test_transform_navigator_family_with_litigation_corpus_type(
-    navigator_family_with_litigation_corpus_type: Identified[NavigatorFamily],
-):
-    result = transform_navigator_family(navigator_family_with_litigation_corpus_type)
-    expected_document_from_family = Document(
-        id="family",
-        title="Litigation family",
-        description="Family summary",
-        labels=[
-            LabelRelationship(
-                type="status",
-                value=Label(
-                    type="status",
-                    id="status::Principal",
-                    value="Principal",
-                ),
-            ),
-            LabelRelationship(
-                type="provider",
-                value=Label(
-                    type="agent",
-                    id="agent::Sabin Center for Climate Change Law",
-                    value="Sabin Center for Climate Change Law",
-                    attributes={
-                        "attribution_url": "testurl.org",
-                        "corpus_text": "Test corpus",
-                        "corpus_image_url": "",
-                    },
-                ),
-            ),
-            LabelRelationship(
-                type="deprecated_category",
-                value=Label(
-                    id="deprecated_category::LITIGATION",
-                    value="LITIGATION",
-                    type="deprecated_category",
-                ),
-            ),
-            LabelRelationship(
-                type="category",
-                value=Label(
-                    id="category::Litigation",
-                    value="Litigation",
-                    type="category",
-                ),
-            ),
-            LabelRelationship(
-                type="activity_status",
-                timestamp=datetime.datetime(2020, 1, 1),
-                value=Label(
-                    type="activity_status",
-                    id="activity_status::Filed",
-                    value="Filed",
-                ),
-            ),
-        ],
-        documents=[
-            DocumentRelationship(
-                type="has_member",
-                value=DocumentWithoutRelationships(
-                    id="document",
-                    title="Litigation family document",
-                    labels=[
-                        LabelRelationship(
-                            type="entity_type",
-                            value=Label(
-                                id="entity_type::Decision",
-                                value="Decision",
-                                type="entity_type",
-                            ),
-                        ),
-                        LabelRelationship(
-                            timestamp=datetime.datetime(2020, 1, 1, 0, 0),
-                            type="activity_status",
-                            value=Label(
-                                attributes={},
-                                documents=[],
-                                id="activity_status::Filed",
-                                labels=[],
-                                type="activity_status",
-                                value="Filed",
-                            ),
-                        ),
-                        LabelRelationship(
-                            type="provider",
-                            value=Label(
-                                type="agent",
-                                id="agent::Sabin Center for Climate Change Law",
-                                value="Sabin Center for Climate Change Law",
-                                attributes={
-                                    "attribution_url": "testurl.org",
-                                    "corpus_text": "Test corpus",
-                                    "corpus_image_url": "",
-                                },
-                            ),
-                        ),
-                        LabelRelationship(
-                            type="category",
-                            value=Label(
-                                id="category::Litigation",
-                                value="Litigation",
-                                type="category",
-                            ),
-                        ),
-                    ],
-                    attributes={
-                        "deprecated_slug": "litigation-document-slug",
-                        "md5_sum": "aaaaa11111bbbbb",
-                        "variant": "Original language",
-                        "status": "PUBLISHED",
-                    },
-                ),
-            ),
-            DocumentRelationship(
-                type="has_member",
-                value=DocumentWithoutRelationships(
-                    id="1.2.3.placeholder",
-                    title="Placeholder litigation family document",
-                    labels=[
-                        LabelRelationship(
-                            type="status",
-                            value=Label(
-                                id="status::Obsolete",
-                                value="Obsolete",
-                                type="status",
-                            ),
-                        ),
-                        LabelRelationship(
-                            type="provider",
-                            value=Label(
-                                type="agent",
-                                id="agent::Sabin Center for Climate Change Law",
-                                value="Sabin Center for Climate Change Law",
-                                attributes={
-                                    "attribution_url": "testurl.org",
-                                    "corpus_text": "Test corpus",
-                                    "corpus_image_url": "",
-                                },
-                            ),
-                        ),
-                        LabelRelationship(
-                            type="category",
-                            value=Label(
-                                id="category::Litigation",
-                                value="Litigation",
-                                type="category",
-                            ),
-                        ),
-                    ],
-                    attributes={
-                        "deprecated_slug": "placeholder-document-slug",
-                        "md5_sum": "aaaaa11111bbbbb",
-                        "status": "PUBLISHED",
-                    },
-                ),
-            ),
-        ],
-        attributes={
-            "deprecated_slug": "litigation-family-slug",
-            "identifier::case_number": "CASE-NUMBER 123",
-            "published_date": "2020-01-0100:00:00Z",
-            "last_updated_date": "2020-01-0100:00:00Z",
-            "status": "PUBLISHED",
-        },
-    )
-    assert_model_list_equality(
-        result.unwrap(),
-        [
-            expected_document_from_family,
-            Document(
-                id="document",
-                title="Litigation family document",
-                labels=[
-                    LabelRelationship(
-                        type="entity_type",
-                        value=Label(
-                            id="entity_type::Decision",
-                            value="Decision",
-                            type="entity_type",
-                        ),
-                    ),
-                    LabelRelationship(
-                        timestamp=datetime.datetime(2020, 1, 1, 0, 0),
-                        type="activity_status",
-                        value=Label(
-                            attributes={},
-                            documents=[],
-                            id="activity_status::Filed",
-                            labels=[],
-                            type="activity_status",
-                            value="Filed",
-                        ),
-                    ),
-                    LabelRelationship(
-                        type="provider",
-                        value=Label(
-                            type="agent",
-                            id="agent::Sabin Center for Climate Change Law",
-                            value="Sabin Center for Climate Change Law",
-                            attributes={
-                                "attribution_url": "testurl.org",
-                                "corpus_text": "Test corpus",
-                                "corpus_image_url": "",
-                            },
-                        ),
-                    ),
-                    LabelRelationship(
-                        type="category",
-                        value=Label(
-                            id="category::Litigation",
-                            value="Litigation",
-                            type="category",
-                        ),
-                    ),
-                ],
-                documents=[
-                    DocumentRelationship(
-                        type="member_of",
-                        value=DocumentWithoutRelationships(
-                            **expected_document_from_family.model_dump()
-                        ),
-                    ),
-                ],
-                attributes={
-                    "deprecated_slug": "litigation-document-slug",
-                    "md5_sum": "aaaaa11111bbbbb",
-                    "variant": "Original language",
-                    "status": "PUBLISHED",
-                },
-            ),
-            Document(
-                id="1.2.3.placeholder",
-                title="Placeholder litigation family document",
-                labels=[
-                    LabelRelationship(
-                        type="status",
-                        value=Label(
-                            id="status::Obsolete",
-                            value="Obsolete",
-                            type="status",
-                        ),
-                    ),
-                    LabelRelationship(
-                        type="provider",
-                        value=Label(
-                            type="agent",
-                            id="agent::Sabin Center for Climate Change Law",
-                            value="Sabin Center for Climate Change Law",
-                            attributes={
-                                "attribution_url": "testurl.org",
-                                "corpus_text": "Test corpus",
-                                "corpus_image_url": "",
-                            },
-                        ),
-                    ),
-                    LabelRelationship(
-                        type="category",
-                        value=Label(
-                            id="category::Litigation",
-                            value="Litigation",
-                            type="category",
-                        ),
-                    ),
-                ],
-                documents=[
-                    DocumentRelationship(
-                        type="member_of",
-                        value=DocumentWithoutRelationships(
-                            **expected_document_from_family.model_dump()
-                        ),
-                    )
-                ],
-                attributes={
-                    "deprecated_slug": "placeholder-document-slug",
-                    "md5_sum": "aaaaa11111bbbbb",
-                    "status": "PUBLISHED",
-                },
-            ),
-        ],
-    )
-
-
-def test_transform_navigator_family_with_litigation_corpus_type_and_litigation_concepts(
-    navigator_family_with_litigation_concepts: Identified[NavigatorFamily],
-):
-    result = transform_navigator_family(navigator_family_with_litigation_concepts)
-    expected_document_from_family = Document(
-        id="family",
-        title="Litigation family",
-        description="Family summary",
-        labels=[
-            LabelRelationship(
-                type="status",
-                value=Label(
-                    type="status",
-                    id="status::Principal",
-                    value="Principal",
-                ),
-            ),
-            LabelRelationship(
-                type="provider",
-                value=Label(
-                    type="agent",
-                    id="agent::Sabin Center for Climate Change Law",
-                    value="Sabin Center for Climate Change Law",
-                    attributes={
-                        "attribution_url": "testurl.org",
-                        "corpus_text": "Test corpus",
-                        "corpus_image_url": "",
-                    },
-                ),
-            ),
-            LabelRelationship(
-                type="deprecated_category",
-                value=Label(
-                    id="deprecated_category::LITIGATION",
-                    value="LITIGATION",
-                    type="deprecated_category",
-                ),
-            ),
-            LabelRelationship(
-                type="category",
-                value=Label(
-                    id="category::Litigation",
-                    value="Litigation",
-                    type="category",
-                ),
-            ),
-            LabelRelationship(
-                type="legal_concept",
-                value=LabelWithoutDocumentRelationships(
-                    id="jurisdiction::High Court of Justice",
-                    labels=[
-                        LabelRelationship(
-                            type="subconcept_of",
-                            value=LabelWithoutDocumentRelationships(
-                                id="jurisdiction::England and Wales",
-                                labels=[],
-                                type="jurisdiction",
-                                value="England and Wales",
-                            ),
-                        )
-                    ],
-                    type="jurisdiction",
-                    value="High Court of Justice",
-                ),
-            ),
-            LabelRelationship(
-                type="legal_concept",
-                value=LabelWithoutDocumentRelationships(
-                    id="jurisdiction::High Court of Justice (Administrative Court)",
-                    labels=[
-                        LabelRelationship(
-                            type="subconcept_of",
-                            value=LabelWithoutDocumentRelationships(
-                                id="jurisdiction::High Court of Justice",
-                                labels=[],
-                                type="jurisdiction",
-                                value="High Court of Justice",
-                            ),
-                        )
-                    ],
-                    type="jurisdiction",
-                    value="High Court of Justice (Administrative Court)",
-                ),
-            ),
-            LabelRelationship(
-                type="legal_concept",
-                value=LabelWithoutDocumentRelationships(
-                    id="jurisdiction::England and Wales",
-                    labels=[],
-                    type="jurisdiction",
-                    value="England and Wales",
-                ),
-            ),
-        ],
-        documents=[
-            DocumentRelationship(
-                type="has_member",
-                value=DocumentWithoutRelationships(
-                    id="document",
-                    title="Litigation family document",
-                    labels=[
-                        LabelRelationship(
-                            type="entity_type",
-                            value=Label(
-                                id="entity_type::Decision",
-                                value="Decision",
-                                type="entity_type",
-                            ),
-                        ),
-                        LabelRelationship(
-                            timestamp=datetime.datetime(2020, 1, 1, 0, 0),
-                            type="activity_status",
-                            value=Label(
-                                attributes={},
-                                documents=[],
-                                id="activity_status::Filed",
-                                labels=[],
-                                type="activity_status",
-                                value="Filed",
-                            ),
-                        ),
-                        LabelRelationship(
-                            type="provider",
-                            value=Label(
-                                type="agent",
-                                id="agent::Sabin Center for Climate Change Law",
-                                value="Sabin Center for Climate Change Law",
-                                attributes={
-                                    "attribution_url": "testurl.org",
-                                    "corpus_text": "Test corpus",
-                                    "corpus_image_url": "",
-                                },
-                            ),
-                        ),
-                        LabelRelationship(
-                            type="category",
-                            value=Label(
-                                id="category::Litigation",
-                                value="Litigation",
-                                type="category",
-                            ),
-                        ),
-                    ],
-                    attributes={
-                        "deprecated_slug": "litigation-document-slug",
-                        "md5_sum": "aaaaa11111bbbbb",
-                        "variant": "Original language",
-                        "status": "PUBLISHED",
-                    },
-                ),
-            ),
-        ],
-        attributes={
-            "deprecated_slug": "litigation-family-slug",
-            "published_date": "2020-01-0100:00:00Z",
-            "last_updated_date": "2020-01-0100:00:00Z",
-            "status": "PUBLISHED",
-        },
-    )
-    assert_model_list_equality(
-        result.unwrap(),
-        [
-            expected_document_from_family,
-            Document(
-                id="document",
-                title="Litigation family document",
-                labels=[
-                    LabelRelationship(
-                        type="entity_type",
-                        value=Label(
-                            id="entity_type::Decision",
-                            value="Decision",
-                            type="entity_type",
-                        ),
-                    ),
-                    LabelRelationship(
-                        timestamp=datetime.datetime(2020, 1, 1, 0, 0),
-                        type="activity_status",
-                        value=Label(
-                            attributes={},
-                            documents=[],
-                            id="activity_status::Filed",
-                            labels=[],
-                            type="activity_status",
-                            value="Filed",
-                        ),
-                    ),
-                    LabelRelationship(
-                        type="provider",
-                        value=Label(
-                            type="agent",
-                            id="agent::Sabin Center for Climate Change Law",
-                            value="Sabin Center for Climate Change Law",
-                            attributes={
-                                "attribution_url": "testurl.org",
-                                "corpus_text": "Test corpus",
-                                "corpus_image_url": "",
-                            },
-                        ),
-                    ),
-                    LabelRelationship(
-                        type="category",
-                        value=Label(
-                            id="category::Litigation",
-                            value="Litigation",
-                            type="category",
-                        ),
-                    ),
-                ],
-                documents=[
-                    DocumentRelationship(
-                        type="member_of",
-                        value=DocumentWithoutRelationships(
-                            **expected_document_from_family.model_dump()
-                        ),
-                    ),
-                ],
-                attributes={
-                    "deprecated_slug": "litigation-document-slug",
-                    "md5_sum": "aaaaa11111bbbbb",
-                    "variant": "Original language",
-                    "status": "PUBLISHED",
-                },
-            ),
-        ],
-    )
-
-
-def test_transform_navigator_family_with_litigation_concepts_missing_parent_label_returns_failure(
-    navigator_family_with_litigation_concept_missing_parent: Identified[
-        NavigatorFamily
-    ],
-):
-    result = transform_navigator_family(
-        navigator_family_with_litigation_concept_missing_parent
-    )
-
-    failure_exception = result.swap().unwrap()
-    assert isinstance(failure_exception, CouldNotTransform)
-    assert (
-        "Unknown parent label 'Missing Parent Label' in relation 'jurisdiction'. "
-        "See family 'family' for details."
-    ) in str(failure_exception)
-
-
-def test_transform_navigator_family_with_multilateral_climate_fund_project(
-    navigator_family_multilateral_climate_fund_project: Identified[NavigatorFamily],
-):
-    result = transform_navigator_family(
-        navigator_family_multilateral_climate_fund_project
-    )
-    expected_document_from_family = Document(
-        id="family",
-        title="Multilateral climate fund project",
-        description="Family summary",
-        labels=[
-            LabelRelationship(
-                type="status",
-                value=Label(
-                    type="status",
-                    id="status::Principal",
-                    value="Principal",
-                ),
-            ),
-            LabelRelationship(
-                type="entity_type",
-                value=Label(
-                    id="entity_type::Project",
-                    value="Project",
-                    type="entity_type",
-                ),
-            ),
-            LabelRelationship(
-                type="activity_status",
-                value=Label(
-                    id="activity_status::Concept approved",
-                    value="Concept approved",
-                    type="activity_status",
-                ),
-                timestamp=datetime.datetime(2020, 1, 1),
-            ),
-            LabelRelationship(
-                type="activity_status",
-                value=Label(
-                    id="activity_status::Approved",
-                    value="Approved",
-                    type="activity_status",
-                ),
-                timestamp=datetime.datetime(2020, 1, 1),
-            ),
-            LabelRelationship(
-                type="activity_status",
-                value=Label(
-                    id="activity_status::Under implementation",
-                    value="Under implementation",
-                    type="activity_status",
-                ),
-                timestamp=datetime.datetime(2020, 1, 1),
-            ),
-            LabelRelationship(
-                type="activity_status",
-                value=Label(
-                    id="activity_status::Completed",
-                    value="Completed",
-                    type="activity_status",
-                ),
-                timestamp=datetime.datetime(2020, 1, 1),
-            ),
-            LabelRelationship(
-                type="activity_status",
-                value=Label(
-                    id="activity_status::Cancelled",
-                    value="Cancelled",
-                    type="activity_status",
-                ),
-                timestamp=datetime.datetime(2020, 1, 1),
-            ),
-            LabelRelationship(
-                type="provider",
-                value=Label(
-                    type="agent",
-                    id="agent::Adaptation Fund",
-                    value="Adaptation Fund",
-                    attributes={
-                        "attribution_url": "testurl.org",
-                        "corpus_text": "Test corpus",
-                        "corpus_image_url": "",
-                    },
-                ),
-            ),
-            LabelRelationship(
-                type="deprecated_category",
-                value=Label(
-                    id="deprecated_category::MCF",
-                    value="MCF",
-                    type="deprecated_category",
-                ),
-            ),
-            LabelRelationship(
-                type="category",
-                value=Label(
-                    id="category::Multilateral Climate Fund project",
-                    value="Multilateral Climate Fund project",
-                    type="category",
-                ),
-            ),
-            LabelRelationship(
-                type="sector",
-                value=Label(
-                    id="sector::Public Sector",
-                    type="sector",
-                    value="Public Sector",
-                ),
-            ),
-            LabelRelationship(
-                type="project_status",
-                value=Label(
-                    id="project_status::Under Implementation",
-                    type="project_status",
-                    value="Under Implementation",
-                ),
-            ),
-            LabelRelationship(
-                type="implementing_agency",
-                value=Label(
-                    id="implementing_agency::International Bank for Reconstruction",
-                    type="implementing_agency",
-                    value="International Bank for Reconstruction",
-                ),
-            ),
-        ],
-        documents=[
-            DocumentRelationship(
-                type="has_member",
-                value=DocumentWithoutRelationships(
-                    id="document_1",
-                    title="Multilateral climate fund project document",
-                    labels=[
-                        LabelRelationship(
-                            type="provider",
-                            value=Label(
-                                type="agent",
-                                id="agent::Adaptation Fund",
-                                value="Adaptation Fund",
-                                attributes={
-                                    "attribution_url": "testurl.org",
-                                    "corpus_text": "Test corpus",
-                                    "corpus_image_url": "",
-                                },
-                            ),
-                        ),
-                        LabelRelationship(
-                            type="category",
-                            value=Label(
-                                id="category::Multilateral Climate Fund project",
-                                value="Multilateral Climate Fund project",
-                                type="category",
-                            ),
-                        ),
-                        LabelRelationship(
-                            type="entity_type",
-                            value=Label(
-                                id="entity_type::Project",
-                                value="Project",
-                                type="entity_type",
-                            ),
-                        ),
-                        LabelRelationship(
-                            type="language",
-                            value=Label(
-                                id="language::eng",
-                                value="eng",
-                                type="language",
-                            ),
-                        ),
-                    ],
-                    attributes={
-                        "deprecated_slug": "document-1-slug",
-                        "variant": "Original language",
-                        "md5_sum": "aaaaa11111bbbbb",
-                        "status": "PUBLISHED",
-                    },
-                ),
-            ),
-            DocumentRelationship(
-                type="has_version",
-                value=DocumentWithoutRelationships(
-                    id="document_2",
-                    title="Project document",
-                    labels=[
-                        LabelRelationship(
-                            type="provider",
-                            value=Label(
-                                type="agent",
-                                id="agent::Adaptation Fund",
-                                value="Adaptation Fund",
-                                attributes={
-                                    "attribution_url": "testurl.org",
-                                    "corpus_text": "Test corpus",
-                                    "corpus_image_url": "",
-                                },
-                            ),
-                        ),
-                        LabelRelationship(
-                            type="category",
-                            value=Label(
-                                id="category::Multilateral Climate Fund project",
-                                value="Multilateral Climate Fund project",
-                                type="category",
-                            ),
-                        ),
-                        LabelRelationship(
-                            type="entity_type",
-                            value=Label(
-                                id="entity_type::Project",
-                                value="Project",
-                                type="entity_type",
-                            ),
-                        ),
-                        LabelRelationship(
-                            type="language",
-                            value=Label(
-                                id="language::eng",
-                                value="eng",
-                                type="language",
-                            ),
-                        ),
-                    ],
-                    attributes={
-                        "deprecated_slug": "document-2-slug",
-                        "variant": "Original language",
-                        "md5_sum": "aaaaa11111bbbbb",
-                        "status": "PUBLISHED",
-                    },
-                ),
-            ),
-        ],
-        attributes={
-            "deprecated_slug": "mcf-family-slug",
-            "identifier::project_id": "XACTMK002A",
-            "identifier::project_approved_ref": "XACTMK002A",
-            "project_co_financing_usd": 100000,
-            "project_fund_spend_usd": 250000,
-            "project_url": "https://www.cif.org/projects",
-            "status": "PUBLISHED",
-        },
-    )
-    assert_model_list_equality(
-        result.unwrap(),
-        [
-            expected_document_from_family,
-            Document(
-                id="document_1",
-                title="Multilateral climate fund project document",
-                labels=[
-                    LabelRelationship(
-                        type="provider",
-                        value=Label(
-                            type="agent",
-                            id="agent::Adaptation Fund",
-                            value="Adaptation Fund",
-                            attributes={
-                                "attribution_url": "testurl.org",
-                                "corpus_text": "Test corpus",
-                                "corpus_image_url": "",
-                            },
-                        ),
-                    ),
-                    LabelRelationship(
-                        type="category",
-                        value=Label(
-                            id="category::Multilateral Climate Fund project",
-                            value="Multilateral Climate Fund project",
-                            type="category",
-                        ),
-                    ),
-                    LabelRelationship(
-                        type="entity_type",
-                        value=Label(
-                            id="entity_type::Project",
-                            value="Project",
-                            type="entity_type",
-                        ),
-                    ),
-                    LabelRelationship(
-                        type="language",
-                        value=Label(
-                            id="language::eng",
-                            value="eng",
-                            type="language",
-                        ),
-                    ),
-                ],
-                documents=[
-                    DocumentRelationship(
-                        type="member_of",
-                        value=DocumentWithoutRelationships(
-                            **expected_document_from_family.model_dump()
-                        ),
-                    )
-                ],
-                attributes={
-                    "deprecated_slug": "document-1-slug",
-                    "variant": "Original language",
-                    "md5_sum": "aaaaa11111bbbbb",
-                    "status": "PUBLISHED",
-                },
-            ),
-            Document(
-                id="document_2",
-                title="Project document",
-                labels=[
-                    LabelRelationship(
-                        type="provider",
-                        value=Label(
-                            type="agent",
-                            id="agent::Adaptation Fund",
-                            value="Adaptation Fund",
-                            attributes={
-                                "attribution_url": "testurl.org",
-                                "corpus_text": "Test corpus",
-                                "corpus_image_url": "",
-                            },
-                        ),
-                    ),
-                    LabelRelationship(
-                        type="category",
-                        value=Label(
-                            id="category::Multilateral Climate Fund project",
-                            value="Multilateral Climate Fund project",
-                            type="category",
-                        ),
-                    ),
-                    LabelRelationship(
-                        type="entity_type",
-                        value=Label(
-                            id="entity_type::Project",
-                            value="Project",
-                            type="entity_type",
-                        ),
-                    ),
-                    LabelRelationship(
-                        type="language",
-                        value=Label(
-                            id="language::eng",
-                            value="eng",
-                            type="language",
-                        ),
-                    ),
-                ],
-                documents=[
-                    DocumentRelationship(
-                        type="is_version_of",
-                        value=DocumentWithoutRelationships(
-                            **expected_document_from_family.model_dump()
-                        ),
-                    )
-                ],
-                attributes={
-                    "deprecated_slug": "document-2-slug",
-                    "variant": "Original language",
-                    "md5_sum": "aaaaa11111bbbbb",
-                    "status": "PUBLISHED",
-                },
-            ),
-        ],
-    )
-
-
 def test_transform_navigator_family_with_published_and_unpublished_documents():
     navigator_family_with_different_document_statuses = Identified(
         id="family",
@@ -2232,7 +905,7 @@ def test_transform_navigator_family_with_published_and_unpublished_documents():
             category="LEGISLATIVE",
             published_date=None,
             last_updated_date=None,
-            corpus=_cclw_laws_and_policies_corpus(),
+            corpus=_cclw_corpus(),
             documents=[
                 NavigatorDocumentFactory.build(
                     import_id="document1",
@@ -2656,7 +1329,7 @@ def test_transform_navigator_family_with_no_published_documents():
             category="LEGISLATIVE",
             published_date=None,
             last_updated_date=None,
-            corpus=_cclw_laws_and_policies_corpus(),
+            corpus=_cclw_corpus(),
             documents=[
                 NavigatorDocumentFactory.build(
                     import_id="document1",
@@ -2899,75 +1572,6 @@ def test_transform_navigator_family_with_no_published_documents():
                     "status": "DELETED",
                 },
             ),
-        ],
-    )
-
-
-def test_transform_navigator_family_with_multilateral_climate_fund_guidance(
-    navigator_family_multilateral_climate_fund_guidance: Identified[NavigatorFamily],
-):
-    result = transform_navigator_family(
-        navigator_family_multilateral_climate_fund_guidance
-    )
-    assert_model_list_equality(
-        result.unwrap(),
-        [
-            Document(
-                id="family",
-                title="Multilateral climate fund guidance",
-                description="Family summary",
-                labels=[
-                    LabelRelationship(
-                        type="status",
-                        value=Label(
-                            type="status",
-                            id="status::Principal",
-                            value="Principal",
-                        ),
-                    ),
-                    LabelRelationship(
-                        type="entity_type",
-                        value=Label(
-                            id="entity_type::Guidance",
-                            value="Guidance",
-                            type="entity_type",
-                        ),
-                    ),
-                    LabelRelationship(
-                        type="provider",
-                        value=Label(
-                            type="agent",
-                            id="agent::Adaptation Fund",
-                            value="Adaptation Fund",
-                            attributes={
-                                "attribution_url": "testurl.org",
-                                "corpus_text": "Test corpus",
-                                "corpus_image_url": "",
-                            },
-                        ),
-                    ),
-                    LabelRelationship(
-                        type="deprecated_category",
-                        value=Label(
-                            id="deprecated_category::REPORTS",
-                            value="REPORTS",
-                            type="deprecated_category",
-                        ),
-                    ),
-                    LabelRelationship(
-                        type="category",
-                        value=Label(
-                            id="category::Multilateral Climate Fund project",
-                            value="Multilateral Climate Fund project",
-                            type="category",
-                        ),
-                    ),
-                ],
-                documents=[],
-                attributes={
-                    "deprecated_slug": "mcf-guidance-family-slug",
-                },
-            )
         ],
     )
